Move display list register parsing to gx_displaylist
This applies to any GX display list, and isn't specific to MDL0 entries.
